update// reformated datastore and changed storage chatmodel to store with datastore
