+++ src/DataStore/DataStore.cpp
@@ -1,0 +1,11 @@
+#include "DataStore.h"
+
+SietchDataStore* DataStore::getSietchDataStore()
+{
+    return SietchDataStore::getInstance();
+}
+
+ChatDataStore* DataStore::getChatDataStore()
+{
+    return ChatDataStore::getInstance();
+}
