+++ src/DataStore/SietchDataStore.h
@@ -15,17 +15,7 @@ class SietchDataStore
         }
 
     public:
-        static SietchDataStore* getInstance()
-        {
-            if(!SietchDataStore::instanced)
-            {
-                SietchDataStore::instanced = true;
-                SietchDataStore::instance = new SietchDataStore();
-            }
-                
-            return SietchDataStore::instance;
-        }
-
+        static SietchDataStore* getInstance();
         void clear();
         void setData(QString key, QString value);
         QString getData(QString key);
@@ -37,28 +27,5 @@ class SietchDataStore
             SietchDataStore::instance = nullptr;
         }
 };
-
-void SietchDataStore::clear()
-{
-    this->data.clear();
-}
-
-void SietchDataStore::setData(QString key, QString value)
-{
-    this->data[key] = value;
-}
-
-QString SietchDataStore::getData(QString key)
-{
-    return this->data[key];
-}
-
-QString SietchDataStore::dump()
-{
-    return "";
-}
-
-SietchDataStore* SietchDataStore::instance = nullptr;
-bool SietchDataStore::instanced = false;
 
 #endif
